requirements.txt
psycopg2
binary
pymongo[srv]
python-dotenv
ipykernel
pdfplumber
python-docx
autogen-agentchat==0.7.1
autogen-ext[openai]
autogen-ext[ollama]
autogen-core
autogenstudio
tiktoken
pytest
streamlit
pymupdf
markdown
autogen-ext[mem0]
autogen-ext[docker]